Backend (#583)
* add backend

* backend

* clean up

* fix

* loss function to backend

* predict and get_device

* add tensorflow backend

* update tensorflow backend

* modify tensorflow backend model_trainer

* update tf backend model

* modify tf backend loss function

* modify tf backend loss function

* update tf backend classification loss function

* update tf backend classification loss function

* update tf backend keras metric

* update test tf backend

* modify nn metric to support multiple backend

* add backend test

* add comma in setup

* comment backend testing

* upgrade tensorflow version

* merge master
