+++ autokeras/constant.py
@@ -4,14 +4,13 @@ GoogleDriveFile = namedtuple('GoogleDriveFile', ['google_drive_id', 'local_name'
 
 
 class Constant:
+    BACKEND = 'torch'
     # Data
-
     VALIDATION_SET_SIZE = 0.08333
     CUTOUT_HOLES = 1
     CUTOUT_RATIO = 0.5
 
     # Searcher
-
     MAX_MODEL_NUM = 1000
     BETA = 2.576
     KERNEL_LAMBDA = 1.0
