+++ autokeras/text/text_supervised.py
@@ -21,15 +21,14 @@ import numpy as np
 import os
 import torch
 
-from autokeras.nn.loss_function import classification_loss
+from autokeras.backend import Backend
 from autokeras.nn.metric import Accuracy
-from autokeras.nn.model_trainer import BERTTrainer
+from autokeras.backend.torch.model_trainer import BERTTrainer
 from autokeras.supervised import SingleModelSupervised
 from autokeras.text.pretrained_bert.utils import PYTORCH_PRETRAINED_BERT_CACHE
 from autokeras.text.pretrained_bert.modeling import BertForSequenceClassification
 from autokeras.text.pretrained_bert.utils import convert_examples_to_features
 from autokeras.text.pretrained_bert.tokenization import BertTokenizer
-from autokeras.utils import get_device
 from torch.utils.data import TensorDataset, DataLoader, SequentialSampler
 
 
@@ -37,7 +36,7 @@ class TextSupervised(SingleModelSupervised, ABC):
 
     def __init__(self, verbose, **kwargs):
         super().__init__(verbose=verbose, **kwargs)
-        self.device = get_device()
+        self.device = Backend.get_device()
 
         # BERT specific
         self.bert_model = 'bert-base-uncased'
@@ -98,7 +97,8 @@ class TextClassifier(TextSupervised):
             verbose: Mode of verbosity.
         """
         super().__init__(verbose=verbose, **kwargs)
-        self.device = get_device()
+        self.device = Backend.get_device()
+        self.verbose = verbose
 
         # BERT specific
         self.bert_model = 'bert-base-uncased'
@@ -182,7 +182,7 @@ class TextClassifier(TextSupervised):
 
     @property
     def loss(self):
-        return classification_loss
+        return Backend.classification_loss
 
     def preprocess(self, x):
         """ Preprocess text data.
